BUG: PR 21897 revisions
* Based on reviewer feedback, add a regression test
that enforces 0-D NumPy integer array support
for `directed_hausdorff` `seed` argument, since this
was previously supported.

* Apply a source code fix for the above issue.

[skip cirrus]

+++ scipy/spatial/tests/test_hausdorff.py
@@ -127,7 +127,7 @@ class TestHausdorff:
         # the two cases from gh-11332
         ([(0,0)],
          [(0,1), (0,0)],
-         0,
+         np.int64(0),
          (0.0, 0, 1)),
         ([(0,0)],
          [(0,1), (0,0)],
